Transfer learning with pretrained Keras model Xception
In this notebook, we used a pre-trained Xception model to predict images in Plant Seedlings Classification competition.

In order to use the pre-trained weights for the Xception model, need to add a new dataset containing the weights. Go to the Data tab and click on 'Add Data'. Then search for the 'Keras Pretrained Model' dataset and add it. Also need to add input data, for that click on 'Add Data'. Go to the competitions tab and then search for the 'Plant Seedlings Classification' dataset and add it.