Plant Seedling Classification using Convolutional Neural Networks
First need to add input data, for that click on 'Add Data'. Go to the competitions tab and then search for the 'Plant Seedlings Classification' dataset and add it. This dataset contains a training set and a testing set of images of plant seedlings at various growth stages. Each image has its own unique ID. The dataset has 12 main plant species which we need to classify the testing set into. For this task, first, we processed and cleaned the data using image processing techniques. Then we build a model and evaluate it